+++ nc_py_api/ex_app/__init__.py
@@ -8,6 +8,6 @@ from .integration_fastapi import (
     set_handlers,
     talk_bot_app,
 )
-from .misc import persistent_storage, verify_version
+from .misc import get_model_path, persistent_storage, verify_version
 from .ui.files_actions import UiActionFileInfo
 from .uvicorn_fastapi import run_app