+++ nc_py_api/ex_app/providers/speech_to_text.py
@@ -1,8 +1,9 @@
 """Nextcloud API for declaring SpeechToText provider."""
 
+import contextlib
 import dataclasses
 
-from ..._exceptions import NextcloudExceptionNotFound
+from ..._exceptions import NextcloudException, NextcloudExceptionNotFound
 from ..._misc import require_capabilities
 from ..._session import AsyncNcSessionApp, NcSessionApp
 
@@ -70,6 +71,16 @@ class _SpeechToTextProviderAPI:
         except NextcloudExceptionNotFound:
             return None
 
+    def report_result(self, task_id: int, result: str = "", error: str = "") -> None:
+        """Report results of speech to text task to Nextcloud."""
+        require_capabilities("app_api", self._session.capabilities)
+        with contextlib.suppress(NextcloudException):
+            self._session.ocs(
+                "PUT",
+                f"{self._session.ae_url}/{self._ep_suffix}",
+                json={"taskId": task_id, "result": result, "error": error},
+            )
+
 
 class _AsyncSpeechToTextProviderAPI:
     """API for registering Speech2Text providers."""
@@ -107,3 +118,13 @@ class _AsyncSpeechToTextProviderAPI:
             )
         except NextcloudExceptionNotFound:
             return None
+
+    async def report_result(self, task_id: int, result: str = "", error: str = "") -> None:
+        """Report results of speech to text task to Nextcloud."""
+        require_capabilities("app_api", await self._session.capabilities)
+        with contextlib.suppress(NextcloudException):
+            await self._session.ocs(
+                "PUT",
+                f"{self._session.ae_url}/{self._ep_suffix}",
+                json={"taskId": task_id, "result": result, "error": error},
+            )