+++ tests/actual_tests/speech2text_provider_test.py
@@ -54,3 +54,14 @@ async def test_speech2text_provider_async(anc_app):
     await anc_app.providers.speech_to_text.unregister(result2.name, not_fail=False)
     assert await anc_app.providers.speech_to_text.get_entry(result2.name) is None
     assert str(result).find("name=") != -1
+
+
+@pytest.mark.require_nc(major=29)
+def test_speech2text_provider_fail_report(nc_app):
+    nc_app.providers.speech_to_text.report_result(999999)
+
+
+@pytest.mark.asyncio(scope="session")
+@pytest.mark.require_nc(major=29)
+async def test_speech2text_provider_fail_report_async(anc_app):
+    await anc_app.providers.speech_to_text.report_result(999999)